src/story26/RecruiterAssistantChatbot.java
package story26;

import java.util.ArrayList;
import java.util.List;
import java.util.Scanner;

public class RecruiterAssistantChatbot {

    // Assume a VirtualJobFairCoordinator class for organizing and promoting virtual job fairs
    static class VirtualJobFairCoordinator {
        private List<String> virtualJobFairDetails = new ArrayList<>();

        public VirtualJobFairCoordinator() {
            // Simulate virtual job fair details (replace with actual data)
            virtualJobFairDetails.add("Virtual Job Fair on [Date] from [Time].");
            virtualJobFairDetails.add("Explore opportunities from leading companies.");
            virtualJobFairDetails.add("Connect with recruiters through virtual booths.");
            // Add more details about the virtual job fair
        }

        public void organizeAndPromoteVirtualJobFair() {
            // Simulate organizing and promoting a virtual job fair
            System.out.println("Virtual Job Fair Details:");
            for (String detail : virtualJobFairDetails) {
                System.out.println("- " + detail);
            }

            // Simulate promoting the virtual job fair to attract a diverse pool of candidates
            System.out.println("\nPromoting the Virtual Job Fair:");
            System.out.println("Share the event on social media platforms.");
            System.out.println("Collaborate with diversity-focused organizations.");
            System.out.println("Highlight inclusive hiring practices in promotional materials.");
            // Add more promotional strategies
        }
    }

    public static void main(String[] args) {
        Scanner scanner = new Scanner(System.in);

        System.out.println("Welcome to the Recruiter Assistant Chatbot!");

        // Assume a VirtualJobFairCoordinator instance for organizing and promoting virtual job fairs
        VirtualJobFairCoordinator jobFairCoordinator = new VirtualJobFairCoordinator();

        do {
            System.out.println("As a recruiter, how can I assist you with organizing a virtual job fair?");
            System.out.println("1. Organize and promote a virtual job fair.");
            System.out.println("2. Exit");

            int userChoice = Integer.parseInt(scanner.nextLine());

            switch (userChoice) {
                case 1:
                    // Organize and promote a virtual job fair
                    jobFairCoordinator.organizeAndPromoteVirtualJobFair();
                    break;
                case 2:
                    System.out.println("Exiting Recruiter Assistant Chatbot. Goodbye!");
                    scanner.close();
                    System.exit(0);
                    break;
                default:
                    System.out.println("Invalid choice. Please enter a valid option.");
                    break;
            }

            System.out.println("Do you want assistance with anything else? (yes/no)");
        } while (scanner.nextLine().equalsIgnoreCase("yes"));

        System.out.println("Exiting Recruiter Assistant Chatbot. Goodbye!");
        scanner.close();
    }
}
